Migrate the ldst CLI to click
Closes #23
